web: split SidebarItem into its own file
