+++ web/src/HUD.tsx
@@ -3,7 +3,8 @@ import AppController from "./AppController"
 import NoMatch from "./NoMatch"
 import Sidebar from "./Sidebar"
 import SidebarAccount from "./SidebarAccount"
-import SidebarResources, { SidebarItem } from "./SidebarResources"
+import SidebarItem from "./SidebarItem"
+import SidebarResources from "./SidebarResources"
 import Statusbar, { StatusItem } from "./Statusbar"
 import LogPane from "./LogPane"
 import HeroScreen from "./HeroScreen"
