+++ web/src/Sidebar.test.tsx
@@ -2,7 +2,8 @@ import React from "react"
 import { mount } from "enzyme"
 import renderer from "react-test-renderer"
 import { MemoryRouter } from "react-router"
-import Sidebar, { SidebarItem } from "./SidebarResources"
+import Sidebar from "./SidebarResources"
+import SidebarItem from "./SidebarItem"
 import { oneResource, twoResourceView } from "./testdata"
 import { ResourceView, TriggerMode } from "./types"
 import PathBuilder from "./PathBuilder"
