+++ web/src/SidebarResources.tsx
@@ -1,14 +1,12 @@
 import React, { PureComponent } from "react"
 import { Link } from "react-router-dom"
-import { combinedStatus } from "./status"
-import { ResourceStatus, ResourceView, TriggerMode } from "./types"
+import { ResourceStatus, ResourceView } from "./types"
 import TimeAgo from "react-timeago"
 import { timeAgoFormatter } from "./timeFormatters"
 import { isZeroTime } from "./time"
 import PathBuilder from "./PathBuilder"
 import SidebarIcon from "./SidebarIcon"
 import SidebarTriggerButton from "./SidebarTriggerButton"
-import { numberOfAlerts } from "./alerts"
 import styled, { keyframes } from "styled-components"
 import { formatBuildDuration } from "./format"
 import {
@@ -21,8 +19,7 @@ import {
   SizeUnit,
   Width,
 } from "./style-helpers"
-
-const moment = require("moment")
+import SidebarItem from "./SidebarItem"
 
 // Styles:
 const barberpole = keyframes`
@@ -132,47 +129,6 @@ let SidebarItemTimeAgo = styled.span`
 type Resource = Proto.webviewResource
 type Build = Proto.webviewBuildRecord
 
-class SidebarItem {
-  name: string
-  isTiltfile: boolean
-  status: ResourceStatus
-  hasEndpoints: boolean
-  lastBuildDur: moment.Duration | null
-  lastDeployTime: string
-  pendingBuildSince: string
-  currentBuildStartTime: string
-  alertCount: number
-  triggerMode: TriggerMode
-  hasPendingChanges: boolean
-  queued: boolean
-  lastBuild: Build | null = null
-
-  /**
-   * Create a pared down SidebarItem from a ResourceView
-   */
-  constructor(res: Resource) {
-    let buildHistory = res.buildHistory || []
-    let lastBuild = buildHistory.length > 0 ? buildHistory[0] : null
-
-    this.name = res.name ?? ""
-    this.isTiltfile = !!res.isTiltfile
-    this.status = combinedStatus(res)
-    this.hasEndpoints = (res.endpointLinks || []).length > 0
-    this.lastBuildDur =
-      lastBuild && lastBuild.startTime && lastBuild.finishTime
-        ? timeDiff(lastBuild.startTime, lastBuild.finishTime)
-        : null
-    this.lastDeployTime = res.lastDeployTime ?? ""
-    this.pendingBuildSince = res.pendingBuildSince ?? ""
-    this.currentBuildStartTime = res.currentBuild?.startTime ?? ""
-    this.alertCount = numberOfAlerts(res)
-    this.triggerMode = res.triggerMode ?? TriggerMode.TriggerModeAuto
-    this.hasPendingChanges = !!res.hasPendingChanges
-    this.queued = !!res.queued
-    this.lastBuild = lastBuild
-  }
-}
-
 type SidebarProps = {
   items: SidebarItem[]
   selected: string
@@ -279,11 +235,4 @@ class SidebarResources extends PureComponent<SidebarProps> {
   }
 }
 
-function timeDiff(start: string, end: string): moment.Duration {
-  let t1 = moment(start)
-  let t2 = moment(end)
-  return moment.duration(t2.diff(t1))
-}
 export default SidebarResources
-
-export { SidebarItem }
