+++ web/src/SidebarTriggerButton.test.tsx
@@ -5,7 +5,8 @@ import SidebarTriggerButton, {
 } from "./SidebarTriggerButton"
 import { ResourceView, TriggerMode } from "./types"
 import { oneResource, twoResourceView } from "./testdata"
-import SidebarResources, { SidebarItem } from "./SidebarResources"
+import SidebarResources from "./SidebarResources"
+import SidebarItem from "./SidebarItem"
 import { MemoryRouter } from "react-router"
 import PathBuilder from "./PathBuilder"
 import fetchMock from "jest-fetch-mock"
